+++ eventum/ui/src/pages/ProjectPage/OutputPluginsTab/OutputPluginParams/HTTPOutputPluginParams.tsx
@@ -302,6 +302,8 @@ export const HTTPOutputPluginParams: FC<HTTPOutputPluginParamsProps> = ({
       <Paper withBorder p="xs">
         <FormatterParams
           value={form.getValues().formatter}
+          errors={form.errors}
+          setErrors={form.setErrors}
           onChange={(values) => form.setFieldValue('formatter', values)}
         />
       </Paper>